requirements.txt
streamlit==1.37.1
pandas==2.2.2
numpy==1.26.4
plotly==5.24.1
python-dateutil==2.9.0.post0
altair==5.3.0
openai>=1.40.0
python-dateutil>=2.9.0.post0
pytz>=2024.1
